% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompleteAuthors.R
\name{format_authors}
\alias{format_authors}
\title{format_authors}
\usage{
format_authors(string)
}
\arguments{
\item{string}{a character vector of names}
}
\description{
This function converts first and middle names to initials
}
