% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_publications.R
\name{format_publications}
\alias{format_publications}
\title{format_publications}
\usage{
format_publications(scholar.profile, author.name = NULL)
}
\arguments{
\item{scholar.profile}{scholar profile ID}

\item{author.name}{name of author to be highlighted using bold font}
}
\value{
a vector of formated publications
}
\description{
Format publication list
}
\examples{
\dontrun{
 library(scholar)
 format_publications("DO5oG40AAAAJ")    
}
}
\author{
R Thériault and modified by Guangchuang Yu
}
