% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_article_cite_history}
\alias{get_article_cite_history}
\title{Gets the citation history of a single article}
\usage{
get_article_cite_history(id, article)
}
\arguments{
\item{id}{a character string giving the id of the scholar}

\item{article}{a character string giving the article id.}
}
\value{
a data frame giving the year, citations per year, and
publication id
}
\description{
Gets the citation history of a single article
}

