% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.r
\name{get_article_scholar_url}
\alias{get_article_scholar_url}
\title{Gets the URL to the google scholar website of an article.}
\usage{
get_article_scholar_url(id, pubid)
}
\arguments{
\item{id}{a character string specifying the Google Scholar ID.}

\item{pubid}{a character string specifying the article id.}
}
\value{
a String that contains the URL to the scholar website of the article
}
\description{
Gets the URL to the google scholar website of an article.
}
