% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthorIds.R
\name{get_author_ids}
\alias{get_author_ids}
\title{Search author by name and return proper Google Scholar IDs}
\usage{
get_author_ids(name = "Roberto Burioni")
}
\arguments{
\item{name}{author name}
}
\value{
a character vector with Google Scholar IDs for the given author name
}
\description{
Search author by name and return proper Google Scholar IDs
}
\examples{
{
   ## Gets profiles of some famous scientist
   get_citation_history(get_author_ids("Roberto Burioni"))
}
}
\author{
Paolo Sonego
}
