% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholar.r
\name{get_citation_history}
\alias{get_citation_history}
\title{Get historical citation data for a scholar}
\usage{
get_citation_history(id)
}
\arguments{
\item{id}{a character string specifying the Google Scholar ID.  If
multiple ids are specified, only the first value is used and a
warning is generated.}
}
\value{
a data frame giving the number of citations per year to
work by the given scholar
}
\description{
Gets the number of citations to a scholar's articles over the past
nine years.
}
\details{
This information is displayed as a bar plot at the top of
a standard Google Scholar page and only covers the past nine
years.
}
