% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{get_citation_year}
\alias{get_citation_year}
\title{Get number of citations in a year}
\usage{
get_citation_year(cid, year)
}
\arguments{
\item{cid}{}

\item{year}{}
}
\description{
Get number of citations in a year
}
