% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{get_citations_per_year}
\alias{get_citations_per_year}
\title{Get citations per year}
\usage{
get_citations_per_year(cid, year)
}
\arguments{
\item{cid}{Citation id of the publication}

\item{year}{Year of publication}
}
\value{
Dataframe with number of cites per year
}
\description{
Gets the number of citations of a publication per year.
}
