% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompleteAuthors.R
\name{get_complete_authors}
\alias{get_complete_authors}
\title{Get the Complete list of authors for a Publication}
\usage{
get_complete_authors(id, pubid, delay = 0.4, initials = TRUE)
}
\arguments{
\item{id}{a Google Scholar ID}

\item{pubid}{a Publication ID from a given google Scholar ID}

\item{delay}{average delay between requests. A delay is needed to stop Google identifying you as a bot}

\item{initials}{if TRUE (default), first and middle names will be abbreviated}
}
\value{
a string containing the complete list of authors
}
\description{
Found as Muhammad Qasim Pasta's solution here https://github.com/jkeirstead/scholar/issues/21
}
\author{
Muhammad Qasim Pasta

Abram B. Fleishman

James H. Conigrave
}
