% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.r
\name{get_publication_data_extended}
\alias{get_publication_data_extended}
\title{Gets the full data for a publication}
\usage{
get_publication_data_extended(id, pub_id, flush = FALSE)
}
\arguments{
\item{id}{a character string specifying the Google Scholar ID.}

\item{pub_id}{a character string specifying the publication id.}

\item{flush}{Whether or not to clear the cache}
}
\value{
a list that contains the full data
}
\description{
Gets the full data for a publication
}
