% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.r
\name{get_publication_date}
\alias{get_publication_date}
\title{Gets the full date for a publication}
\usage{
get_publication_date(id, pub_id, flush = FALSE)
}
\arguments{
\item{id}{a character string specifying the Google Scholar ID.}

\item{pub_id}{a character string specifying the publication id.}

\item{flush}{Whether or not to clear the cache}
}
\value{
a String that contains the publication date
}
\description{
Gets the full date for a publication
}
