% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.r
\name{get_publications_all_authors}
\alias{get_publications_all_authors}
\title{Gets publications with complete author lists}
\usage{
get_publications_all_authors(id, delay = 0.8, initials = TRUE, ...)
}
\arguments{
\item{id}{a character string specifying the Google Scholar ID.}

\item{delay}{average delay between requests for complete author lists. See
\code{\link{get_complete_authors}}.}

\item{initials}{if TRUE, first and middle names in completed author lists
will be abbreviated. See \code{\link{get_complete_authors}}.}

\item{...}{additional arguments passed to \code{\link{get_publications}}.}
}
\value{
a data frame listing the publications and their details, with
truncated author lists replaced where possible.
}
\description{
Gets a scholar's publications and fills truncated author lists by calling
\code{\link{get_complete_authors}} only for publications whose author field
contains \code{...}.
}
