% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholar.r
\name{get_scholar_id}
\alias{get_scholar_id}
\title{Search for Google Scholar ID by name and affiliation}
\usage{
get_scholar_id(last_name = "", first_name = "", affiliation = NA)
}
\arguments{
\item{last_name}{Researcher last name.}

\item{first_name}{Researcher first name.}

\item{affiliation}{Researcher affiliation.}
}
\value{
Google Scholar ID as a character string.
}
\description{
Search for Google Scholar ID by name and affiliation
}
\examples{
get_scholar_id(first_name = "kristopher", last_name = "mcneill")
\donttest{
get_scholar_id(first_name = "michael", last_name = "sander", affiliation = NA)
get_scholar_id(first_name = "michael", last_name = "sander", affiliation = "eth")
get_scholar_id(first_name = "michael", last_name = "sander", affiliation = "ETH Zurich")
get_scholar_id(first_name = "michael", last_name = "sander", affiliation = "Mines")
get_scholar_id(first_name = "james", last_name = "babler")
}
}
