% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.r
\name{get_scholar_metrics}
\alias{get_scholar_metrics}
\title{Calculate citation metrics for a scholar}
\usage{
get_scholar_metrics(id, pagesize = 100, flush = FALSE)
}
\arguments{
\item{id}{a character string specifying the Google Scholar ID}

\item{pagesize}{an integer specifying the number of articles to fetch in one
batch. See \code{\link{get_publications}}.}

\item{flush}{should the publication cache be flushed?}
}
\value{
a data frame with h-index, g-index, i10-index, i50-index, i100-index,
total citations, and publication count
}
\description{
Fetches a scholar's publications with \code{\link{get_publications}} and
calculates common citation metrics.
}
