% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_scholar_resp}
\alias{get_scholar_resp}
\title{Recursively try to GET a Google Scholar Page storing session cookies}
\usage{
get_scholar_resp(url, attempts_left = 5)
}
\arguments{
\item{url}{URL to fetch}

\item{attempts_left}{The number of times to try and fetch the page}
}
\value{
an \code{httr::\link{response}} object
}
\description{
see \code{\link{scholar-package}} documentation for details about Scholar
session cookies.
}
\seealso{
\code{httr::\link{GET}}
}
