% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{search_publication}
\alias{search_publication}
\title{Search publication}
\usage{
search_publication(text, exact_text = "", author = "", from = 0)
}
\arguments{
\item{text}{}

\item{exact_text}{}

\item{author}{}

\item{from}{}
}
\description{
Search publication
}
