% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_scholar_mirror}
\alias{set_scholar_mirror}
\title{set_scholar_mirror}
\usage{
set_scholar_mirror(mirror = "https://g88.i-research.edu.eu.org/ggxs")
}
\arguments{
\item{mirror}{compatible scholar mirror}
}
\description{
set scholar mirror
}
\details{
setting google scholar mirror
}
\author{
Guangchuang Yu
}
