% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.r
\name{compare_scholars}
\alias{compare_scholars}
\title{Compare the citation records of multiple scholars}
\usage{
compare_scholars(ids, pagesize = 100)
}
\arguments{
\item{ids}{a vector of Google Scholar IDs}

\item{pagesize}{an integer specifying the number of articles to
fetch for each scholar}
}
\value{
a data frame giving the ID of each scholar and the total
number of citations received by work published in a year.
}
\description{
Compares the citation records of multiple scholars.  This function
compiles a data frame comparing the citations received by each of
the scholar's publications by year of publication.
}
\examples{
{
\donttest{
    ## How do Richard Feynmann and Stephen Hawking compare?
    ids <- c("B7vSqZsAAAAJ", "qj74uXkAAAAJ")
    df <- compare_scholars(ids)
}
}
}
