% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{get_citation_year}
\alias{get_citation_year}
\title{Get number of citations in a year}
\usage{
get_citation_year(cid, year)
}
\arguments{
\item{cid}{Citation id of the publication}

\item{year}{Year to search}
}
\value{
Number of publications in a year
}
\description{
Gets the number of citations of a publication in a year.
}
