% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.r
\name{get_num_articles}
\alias{get_num_articles}
\title{Calculates how many articles a scholar has published}
\usage{
get_num_articles(id)
}
\arguments{
\item{id}{a character string giving the Google Scholar ID}
}
\value{
an integer value (max 100)
}
\description{
Calculate how many articles a scholar has published.
}
