% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompleteAuthors.R
\name{getCompleteAuthors}
\alias{getCompleteAuthors}
\title{Get the Complete list of authors for a Publication}
\usage{
getCompleteAuthors(id, pubid)
}
\arguments{
\item{id}{a Google Scholar ID}

\item{pubid}{a Publication ID from a giving google Schalar ID}
}
\value{
a string containing the complete list of authors
}
\description{
Found as Muhammad Qasim Pasta's solution here https://github.com/jkeirstead/scholar/issues/21
}
\author{
Muhammad Qasim Pasta

Abram B. Fleishman
}
