% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{search_publication}
\alias{search_publication}
\title{Search publication}
\usage{
search_publication(text, exact_text = "", author = "", from = 0,
  until = 10)
}
\arguments{
\item{text}{Text to search}

\item{exact_text}{Exact text to search}

\item{author}{Author from publication}

\item{from}{Starting point of search}

\item{until}{Ending point of search}
}
\value{
Publication dataframe
}
\description{
Searches publications in Google Scholar.
}
